multreprelauch

+++ src/controllers/Products/testImage.js
@@ -1,6 +1,12 @@
 const cloudinary = require('../../middleware/cloudinary.js')
 const fs = require('fs')
 
+const Product = require("../../models/Product.js");
+const Category = require("../../models/Category.js");
+
+const normalizeString = require('../../utils/normalizeString.js');
+const convertToInt = require("../../utils/convertToInt.js");
+
 const cloudinaryMethod = async (file) =>{
     return new Promise (resolve=>{
         cloudinary.uploader.upload(file,(err,res)=>{
@@ -16,19 +22,37 @@ const cloudinaryMethod = async (file) =>{
 const testInput =async (req,res,next)=>{
     const {input} = req.body
     const formData = JSON.parse(input)
-    console.log(formData)
-    console.log(req.files)
-    
+    const { name, img, price, description, stock, categories } = formData;
+    if (!name || /* !img || */ !price || !description || !stock || !categories) return res.status(400).json({msg:"Check the fields."})
     try {
     const urls = [];
     const files = req.files;
     for(const file of files){
         const {path} = file;
         const newPath = await cloudinaryMethod(path)
-        urls.push(newPath)
+        urls.push(newPath.res)
         fs.unlinkSync(path)
     }
-    res.send({urls,formData})
+    console.log(urls)
+    const newProduct = await Product.create({
+        name: normalizeString(name),
+        price, 
+        img: [...img,...urls],
+        description, 
+        stock: convertToInt(stock),
+        rating: convertToInt("0")
+     });
+
+     categories.forEach(async (item) => {
+        const [newCategory, boolCreate] = await Category.findOrCreate({
+           where: {
+              name: normalizeString(item),
+           }
+        });
+        await newProduct.addCategory(newCategory);
+     })
+
+     res.status(201).json({ msg: "Producto creado exitosamente", name: newProduct.name });
   } catch (error) {
       console.log(error)
   }
